feat: Add Developer menu with clear storage option
Adds a Developer menu to the application menu bar with a "Clear application
storage" option that clears localStorage, sessionStorage, and auth state.
Extracts storage clearing logic into a shared utility function.
