+++ apps/array/src/renderer/lib/clearStorage.ts
@@ -1,0 +1,22 @@
+import { logger } from "./logger";
+
+const log = logger.scope("clear-storage");
+
+export function clearApplicationStorage(): void {
+  const confirmed = window.confirm(
+    "Are you sure you want to clear all application storage?\n\nThis will remove:\n• All registered folders\n• UI state (sidebar preferences, etc.)\n• Task directory mappings\n\nYour files will not be deleted from your computer.",
+  );
+
+  if (confirmed) {
+    window.electronAPI.folders
+      .clearAllData()
+      .then(() => {
+        localStorage.clear();
+        window.location.reload();
+      })
+      .catch((error: unknown) => {
+        log.error("Failed to clear storage:", error);
+        alert("Failed to clear storage. Please try again.");
+      });
+  }
+}
